+++ backend/src/main/java/com/code/theaterapp/stage/dtos/StageDTO.java
@@ -3,6 +3,6 @@ package com.code.theaterapp.stage.dtos;
 public record StageDTO(
         String name,
         Integer capacity,
-        Long venueId
+        Integer venueId
 ) {
 }
